+++ lib/sport_ngin_aws_auditor/cache_instance.rb
@@ -7,40 +7,36 @@ module SportNginAwsAuditor
     extend AWSWrapper
 
     class << self
-      attr_accessor :instances, :reserved_instances, :retired_reserved_instances
-
-      def get_instances(tag_name=nil)
-        return @instances if @instances
+      def get_instances(client, tag_name=nil)
         account_id = get_account_id
-        @instances = cache.describe_cache_clusters.cache_clusters.map do |instance|
+        client.describe_cache_clusters.cache_clusters.map do |instance|
           next unless instance.cache_cluster_status.to_s == 'available'
-          new(instance, account_id, tag_name, cache)
+          new(instance, account_id, tag_name, client)
         end.compact
       end
 
-      def get_reserved_instances
-        return @reserved_instances if @reserved_instances
-        @reserved_instances = cache.describe_reserved_cache_nodes.reserved_cache_nodes.map do |instance|
+      def get_reserved_instances(client)
+        client.describe_reserved_cache_nodes.reserved_cache_nodes.map do |instance|
           next unless instance.state.to_s == 'active'
           new(instance)
         end.compact
       end
 
-      def get_retired_reserved_instances
-        return @retired_reserved_instances if @retired_reserved_instances
-        @retired_reserved_instances = cache.describe_reserved_cache_nodes.reserved_cache_nodes.map do |instance|
+      def get_retired_reserved_instances(client)
+        client.describe_reserved_cache_nodes.reserved_cache_nodes.map do |instance|
           next unless instance.state == 'retired'
           new(instance)
         end.compact
       end
     end
 
-    attr_accessor :id, :name, :instance_type, :scope, :engine, :count, :tag_value, :tag_reason, :expiration_date
-    def initialize(cache_instance, account_id=nil, tag_name=nil, cache=nil)
+    attr_accessor :id, :name, :instance_type, :scope, :engine, :count, :tag_value, :tag_reason, :expiration_date, :availability_zone
+    def initialize(cache_instance, account_id=nil, tag_name=nil, client=nil)
       if cache_instance.class.to_s == "Aws::ElastiCache::Types::ReservedCacheNode"
         self.id = cache_instance.reserved_cache_node_id
         self.name = cache_instance.reserved_cache_node_id
         self.scope = nil
+        self.availability_zone = nil
         self.instance_type = cache_instance.cache_node_type
         self.engine = cache_instance.product_description
         self.count = cache_instance.cache_node_count
@@ -49,6 +45,7 @@ module SportNginAwsAuditor
         self.id = cache_instance.cache_cluster_id
         self.name = cache_instance.cache_cluster_id
         self.scope = nil
+        self.availability_zone = nil
         self.instance_type = cache_instance.cache_node_type
         self.engine = cache_instance.engine
         self.count = cache_instance.num_cache_nodes
@@ -59,7 +56,7 @@ module SportNginAwsAuditor
           arn = "arn:aws:elasticache:#{region}:#{account_id}:cluster:#{self.id}"
 
           # go through to see if the tag we're looking for is one of them
-          cache.list_tags_for_resource(resource_name: arn).tag_list.each do |tag|
+          client.list_tags_for_resource(resource_name: arn).tag_list.each do |tag|
             if tag.key == tag_name
               self.tag_value = tag.value
             elsif tag.key == 'no-reserved-instance-reason'