+++ spec/sport_ngin_aws_auditor/audit_data_spec.rb
@@ -28,75 +28,94 @@ module SportNginAwsAuditor
 
     context '#initialization' do
       it 'should gather instance data' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.selected_audit_type).to eq("instances")
       end
 
       it 'should gather reserved instance data' do
-        audit_results = AuditData.new(false, true, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => true, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.selected_audit_type).to eq("reserved")
       end
 
       it 'should by default gather instance data' do
-        audit_results = AuditData.new(true, true, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => true, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.selected_audit_type).to eq("instances")
       end
 
       it 'should gather all data to compare' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.selected_audit_type).to eq("all")
       end
 
       it 'should use EC2Instance class' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.klass).to eq(SportNginAwsAuditor::EC2Instance)
       end
 
       it 'should use EC2Instance class' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.tag_name).to eq("no-reserved-instance")
+      end
+
+      it 'should use EC2Instance class' do
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
+        expect(audit_results.region).to eq("us-east")
       end
     end
 
     context '#instances?' do
       it 'should return true' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.instances?).to eq(true)
       end
 
       it 'should return true' do
-        audit_results = AuditData.new(false, true, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => true, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.instances?).to eq(false)
       end
     end
 
     context '#reserved?' do
       it 'should return true' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.reserved?).to eq(false)
       end
 
       it 'should return true' do
-        audit_results = AuditData.new(false, true, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => true, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.reserved?).to eq(true)
       end
     end
 
     context '#all?' do
       it 'should return true' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.all?).to eq(false)
       end
 
       it 'should return true' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         expect(audit_results.all?).to eq(true)
       end
     end
 
     context '#gather_data' do
       it 'should gather some empty results by comparison' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         audit_results.gather_data
         expect(audit_results.data).to eq([@instance, @instance])
         expect(audit_results.retired_tags).to eq([])
@@ -104,7 +123,8 @@ module SportNginAwsAuditor
       end
 
       it 'should gather some empty results from just instances' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         audit_results.gather_data
         expect(audit_results.data).to eq([@instance, @instance])
         expect(audit_results.retired_tags).to eq([])
@@ -112,7 +132,8 @@ module SportNginAwsAuditor
       end
 
       it 'should gather some empty results from just reserved' do
-        audit_results = AuditData.new(false, true, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => true, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         audit_results.gather_data
         expect(audit_results.data).to eq([@instance, @instance])
         expect(audit_results.retired_tags).to eq(nil)
@@ -122,7 +143,8 @@ module SportNginAwsAuditor
 
     context '#gather_instances_data' do
       it 'should gather some instances data but not convert' do
-        audit_results = AuditData.new(true, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => true, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         result1, result2 = audit_results.gather_instances_data
         expect(result1).to eq({'instance1' => 1, 'instance2' => 1})
         expect(result2).to eq([])
@@ -131,19 +153,12 @@ module SportNginAwsAuditor
 
     context '#gather_all_data' do
       it 'should gather some comparison data but not convert' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
+        info = {:instances => false, :reserved => false, :class => "EC2Instance", :tag_name => "no-reserved-instance", :regexes => @ignore_instances_regexes, :region => 'us-east-1'}
+        audit_results = AuditData.new(info)
         result1, result2, result3 = audit_results.gather_all_data
         expect(result1).to eq({'instance1' => 1, 'instance2' => 1})
         expect(result2).to eq([])
         expect(result3).to eq(@retired_ris)
-      end
-    end
-
-    context '#gather_region' do
-      it 'should gather the region from an instance' do
-        audit_results = AuditData.new(false, false, "EC2Instance", "no-reserved-instance", @ignore_instances_regexes)
-        audit_results.gather_region(@ec2_instances)
-        expect(audit_results.region).to eq('us-east')
       end
     end
   end